serialize_tags: make .filter(tags__contains='string') work
the example in readme says this example works, but actually it splits
the argument by ',' unconditionally, expecting a list.  so a string will
be made into a "s,t,r,i,n,g".  so we should handle getting passed either
a single string and pass it through unchanged, or a list and join them

actually if a list, we'd want to dedupe so we use a set which is the
native type of the library for tags anyways